backend_python_skeleton/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
sqlalchemy[asyncio]==2.0.21
databases==0.7.4
python-dotenv==1.0.1
passlib[bcrypt]==1.7.5
python-jose==3.3.0
aiosqlite==0.18.0
